Implemented another stopping criteria
The algorithm proceeds until the best solution during the evolution process doesn't change to a better value for a predefined value of generations. This predefined value is 30% of the generation number which the best solution has found so far. I.e. the algorithm reaches to a value of 200 at generation 50, then this value doesn't change for 15 generations (30% of 50), so the algorithm stops.